+++ 1-libft/ft_calloc.c
@@ -6,16 +6,16 @@
 /*   By: sanseo <sanseo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 19:24:35 by sanseo            #+#    #+#             */
-/*   Updated: 2023/03/18 19:31:36 by sanseo           ###   ########.fr       */
+/*   Updated: 2023/03/19 15:47:57 by sanseo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include<stdlib.h>
 
-void	*ft_calloc(unsigned long long count, unsigned long long size)
+void	*ft_calloc(size_t count, size_t size)
 {
-	char				*mem;
-	unsigned long long	i;
+	char	*mem;
+	size_t	i;
 
 	mem = (char *)malloc(count * size);
 	if (mem == NULL)
