+++ 1-libft/ft_memcmp.c
@@ -6,15 +6,17 @@
 /*   By: sanseo <sanseo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 18:19:01 by sanseo            #+#    #+#             */
-/*   Updated: 2023/03/18 20:26:29 by sanseo           ###   ########.fr       */
+/*   Updated: 2023/03/19 15:49:05 by sanseo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned long long n)
+#include<stdlib.h>
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	const char				*str1;
-	const char				*str2;
-	unsigned long long		i;
+	const char	*str1;
+	const char	*str2;
+	size_t		i;
 
 	str1 = (const char *)s1;
 	str2 = (const char *)s2;
