+++ 1-libft/ft_memmove.c
@@ -6,15 +6,17 @@
 /*   By: sanseo <sanseo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/14 17:27:06 by sanseo            #+#    #+#             */
-/*   Updated: 2023/03/18 20:26:46 by sanseo           ###   ########.fr       */
+/*   Updated: 2023/03/19 15:49:46 by sanseo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, unsigned long long n)
+#include<stdlib.h>
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	char				*dst;
-	const char			*sc;
-	unsigned long long	i;
+	char		*dst;
+	const char	*sc;
+	size_t		i;
 
 	dst = (char *)dest;
 	sc = (const char *)src;
