+++ 1-libft/ft_strdup.c
@@ -6,13 +6,13 @@
 /*   By: sanseo <sanseo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 19:31:51 by sanseo            #+#    #+#             */
-/*   Updated: 2023/03/18 20:09:35 by sanseo           ###   ########.fr       */
+/*   Updated: 2023/03/19 15:51:32 by sanseo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include<stdlib.h>
 
-unsigned long long	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	const char	*ptr;
 
@@ -24,8 +24,8 @@ unsigned long long	ft_strlen(const char *s)
 
 char	*ft_strdup(const char *s1)
 {
-	char				*str;
-	unsigned long long	i;
+	char	*str;
+	size_t	i;
 
 	i = ft_strlen(s1);
 	str = (char *)malloc(i);
